Butt Bugfixes
Fixes double messages when farting on someone.

Fixes messages not appearing for players when farting.

Atomic Asses now reset properly upon detonation.

Bluespace Asses may now be researched and built.

Xenomorphs can now be butchered for their ass. Dear god.

And I heard a voice in the midst of the four beasts
And I looked, and behold, a pale horse
And his name, that sat on him, was Death

And farts followed with him.