Syndicate Radio
Grants Syndicate Silicons & Syndicate Radio Implants the ability to speak on syndicomms.

I mean, the implants aren't speaking.
It's the person with the implant.
Implantee? That sounds weird.
Implanted, that's it.

Though, is it actually the implant speaking?
One could say that talking through a radio means that the radio itself is speaking for you.
But that would require stretching it to our mouths talking for us.
I don't like that.