Readds/enables the Mysterious Arrow ruin
About The Pull Request
It reenables a ruin disabled by Bee when it was ported into it. The ruin has 3 syndicate troopers, 4 forbidden/forgotten and one more mob that I currently forgot the name of. Behind all of that is a Mysterious Arrow that gives anyone that uses it a 50/50 of death or a holoparasite. The arrow can also be used on holoparasites to buff them

Why It's Good For The Game
More ruins in lavaland which could always use more depth. More things to Holoparasites giving them more interest.

Changelog
🆑
add: The Mysterious Arrow ruin can now be found once more in Lavaland with a 20% chance!
/🆑