Simplifies the pull request template (#6479)
* Simplifies the pull request template

Simplifies the PR template, removing a lot of blank space in the details section and simplifying an unnecessarilly long comment in the testing photographs section that makes it too long to read, too long to be useful and too long to be easilly deleted.
Removes the part stating to add closing tags in the 'why its good for the game' section since they are placed at the top of the PR rather than in the good for the game section.

'Include any screenshots/videos of the modified code functioning successfully ideally including edge cases.' says everything that needs to be said without too much bloat.

:cl:
code: Simplifies the PR template.
/:cl:

* Update .github/PULL_REQUEST_TEMPLATE.md

* Update .github/PULL_REQUEST_TEMPLATE.md