Contractor Jail Improvements
Reduces the amount of dizziness the contractor victim gets from 85 to 30.

Reduces the amount of confusion the contractor victim gets from 50 to 20.

Reduces the time spent in the timeshare from 4 minutes to 2 minutes.

Gives the contractor victim more healing chems.
Now adds 20u of: Omnizine, Tricordrazine, Salbutamol and Iron. (Originally just Omnizine)

Properly uses SECONDS defines in that section of the code.