Strips down aiming
Aiming no longer creates a radial menu with the surrender/item drop/etc buttons.

This was found to be too easy to powergame as a means to hardstun someone by mistake, so it is now a purely RP connected means of "holding someone up"

The original notifications of the person equipping, dropping, and surrendering  all remain.