Flash/Deafen Updates
Lowers the volume of flashbang.ogg and gives it less of a high pitched focus.

Changes flash_ring.ogg to be less high pitched and more focused on the deafness aspect.

Adds the Deafened and Flashed status effects.

Removes the original fullscreen white effect when flashed, replacing it with a darker vision reduction.

Cleans up a needless proc argument in flash_act.

Being deaf now alerts you when you are deaf.