GAGS & Sprites
Adds GAGS support to the ninja clothing to allow for custom lighting.

Updates the sprites of the ninja suit to the MODsuit versions.

Adds the ninja jumpsuit.

Updates energy costs to accurately reflect that it's a percentage cost of the base cell.

Reduces the Energy Net's power cost

Gives the ninja a dedicated power cell rather than a varedited one.

Adds cell damage and power cell charge to the status check ability.