Flashed Status Effect Changes
Moves the status effect application to be independent of the strength of the flash.
Any brightness that would damage is enough to cause the effect.

Changes the flavor text of the status effect to reflect the examine text.

Changes the duration of the status effect to be more in line with the length of an actual flash stun.